GC.cpp: PyrGc::Collect: promote mNumToScan to u64
Undo previous commit of static_casting as that stopped the crash, but didn't solve the underlying issue.

If this overflows then a different gc strategy will have to be considered.

While fixing the following segfault on linux, mNumToScan overflows as 1810000 function objects must be kept around until the line finishes executing.
An alternative solution would be to try and free those function objects, but this requires a charger change.
```
b = Pwhite.new(0, 10, inf).asStream;
Array.fill(1810000, {b.next});`
```

+++ lang/LangSource/GC.h
@@ -126,7 +126,7 @@ public:
     // users should not call anything below.
 
     void Collect();
-    void Collect(int32 inNumToScan);
+    void Collect(uint64 inNumToScan);
     void LazyCollect() {
         if (mUncollectedAllocations > kLazyCollectThreshold)
             Collect();
@@ -197,7 +197,7 @@ private:
     PyrObjectHdr mGrey;
 
     int32 mPartialScanSlot;
-    int32 mNumToScan;
+    uint64 mNumToScan;
     int32 mNumGrey;
 
     int32 mFlips, mCollects, mAllocTotal, mScans, mNumAllocs, mStackScans, mNumPartialScans, mSlotsScanned,
